Feature/java 17

+++ src/test/java/com/sapher/youtubedl/YoutubeDLTest.java
@@ -34,7 +34,7 @@ public class YoutubeDLTest {
         request.setOption("version");
         YoutubeDLResponse response = YoutubeDL.execute(request);
 
-        int elapsedTime = (int) (System.nanoTime() - startTime);
+        long elapsedTime = System.nanoTime() - startTime;
 
         Assert.assertTrue(elapsedTime > response.getElapsedTime());
     }
@@ -101,4 +101,5 @@ public class YoutubeDLTest {
     public void testFailGetNonExistentVideoInfo() throws YoutubeDLException {
         YoutubeDL.getVideoInfo(NONE_EXISTENT_VIDEO_URL);
     }
+
 }
